requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
stripe==7.8.0
prometheus-client==0.19.0
sentry-sdk==1.39.1
celery==5.3.4
flower==2.0.1
loguru==0.7.2
tenacity==8.2.3
aiofiles==23.2.1
